+++ addons/account/models/__init__.py
@@ -31,6 +31,7 @@ from . import account_cash_rounding
 from . import account_incoterms
 from . import digest
 from . import res_users
+from . import ir_attachment
 from . import ir_actions_report
 from . import res_currency
 from . import account_report
