+++ addons/account_edi/models/account_move.py
@@ -398,22 +398,6 @@ class AccountMove(models.Model):
         return self._get_edi_document(edi_format).attachment_id
 
     ####################################################
-    # Import Electronic Document
-    ####################################################
-
-    def _get_create_document_from_attachment_decoders(self):
-        # OVERRIDE
-        res = super()._get_create_document_from_attachment_decoders()
-        res.append((10, self.env['account.edi.format'].search([])._create_document_from_attachment))
-        return res
-
-    def _get_update_invoice_from_attachment_decoders(self, invoice):
-        # OVERRIDE
-        res = super()._get_update_invoice_from_attachment_decoders(invoice)
-        res.append((10, self.env['account.edi.format'].search([])._update_invoice_from_attachment))
-        return res
-
-    ####################################################
     # Business operations
     ####################################################
