+++ addons/account_edi/tests/test_import_vendor_bill.py
@@ -10,7 +10,7 @@ class TestImportVendorBill(AccountTestInvoicingCommon):
 
         def retrieve_partner(vat, import_vat):
             self.partner_a.with_context(no_vat_validation=True).vat = vat
-            return self.env['account.edi.format']._retrieve_partner(vat=import_vat)
+            return self.env['res.partner']._retrieve_partner(vat=import_vat)
 
         self.assertEqual(self.partner_a, retrieve_partner('BE0477472701', 'BE0477472701'))
         self.assertEqual(self.partner_a, retrieve_partner('BE0477472701', '0477472701'))
