+++ addons/account_edi_ubl_cii/models/account_edi_common.py
@@ -248,25 +248,26 @@ class AccountEdiCommon(models.AbstractModel):
     # Import invoice
     # -------------------------------------------------------------------------
 
-    def _import_invoice(self, journal, filename, tree, existing_invoice=None):
-        move_type, qty_factor = self._get_import_document_amount_sign(filename, tree)
+    def _import_invoice(self, invoice, tree, new=False):
+        # Not able to decode the move_type from the xml.
+        move_type, qty_factor = self._get_import_document_amount_sign(tree)
         if not move_type:
             return
+
+        # Check for inconsistent move_type.
+        journal = invoice.journal_id
         if journal.type == 'sale':
             move_type = 'out_' + move_type
         elif journal.type == 'purchase':
             move_type = 'in_' + move_type
         else:
             return
-        if existing_invoice and existing_invoice.move_type != move_type:
+        if not new and invoice.move_type != move_type:
             return
 
-        with (existing_invoice or self.env['account.move']).with_context(
-            account_predictive_bills_disable_prediction=True,
-            default_move_type=move_type,
-            default_journal_id=journal.id,
-        )._get_edi_creation() as invoice:
-            logs = self._import_fill_invoice_form(journal, tree, invoice, qty_factor)
+        # Update the invoice.
+        invoice.move_type = move_type
+        logs = self._import_fill_invoice_form(invoice, tree, qty_factor)
         if invoice:
             if logs:
                 body = _(
@@ -275,7 +276,7 @@ class AccountEdiCommon(models.AbstractModel):
                 )
             else:
                 body = _("<strong>Format used to import the invoice: %s</strong>", str(self._description))
-            invoice.with_context(no_new_invoice=True).message_post(body=body)
+            invoice.message_post(body=body)
 
         # === Import the embedded PDF in the xml if some are found ===
 
@@ -301,11 +302,11 @@ class AccountEdiCommon(models.AbstractModel):
                     'mimetype': 'application/pdf',
                 })
         if attachments:
-            invoice.with_context(no_new_invoice=True).message_post(attachment_ids=attachments.ids)
+            invoice.message_post(attachment_ids=attachments.ids)
 
-        return invoice
+        return True
 
-    def _import_fill_invoice_allowance_charge(self, tree, invoice, journal, qty_factor):
+    def _import_fill_invoice_allowance_charge(self, tree, invoice, qty_factor):
         logs = []
         if '{urn:oasis:names:specification:ubl:schema:xsd' in tree.tag:
             is_ubl = True
@@ -357,10 +358,10 @@ class AccountEdiCommon(models.AbstractModel):
             tax_ids = []
             for tax_categ_percent_el in allow_el.findall(tax_xpath):
                 tax = self.env['account.tax'].search([
-                    ('company_id', '=', journal.company_id.id),
+                    ('company_id', '=', invoice.company_id.id),
                     ('amount', '=', float(tax_categ_percent_el.text)),
                     ('amount_type', '=', 'percent'),
-                    ('type_tax_use', '=', journal.type),  # Journal type is ensured by _create_invoice_from_xml_tree to be either 'sale' or 'purchase'
+                    ('type_tax_use', '=', invoice.journal_id.type),  # Journal type is ensured by _create_invoice_from_xml_tree to be either 'sale' or 'purchase'
                 ], limit=1)
                 if tax:
                     tax_ids += tax.ids
@@ -563,16 +564,16 @@ class AccountEdiCommon(models.AbstractModel):
             'product_uom_id': product_uom_id,
         }
 
-    def _import_fill_invoice_line_taxes(self, journal, tax_nodes, invoice_line_form, inv_line_vals, logs):
+    def _import_fill_invoice_line_taxes(self, tax_nodes, invoice_line, inv_line_vals, logs):
         # Taxes: all amounts are tax excluded, so first try to fetch price_include=False taxes,
         # if no results, try to fetch the price_include=True taxes. If results, need to adapt the price_unit.
         inv_line_vals['taxes'] = []
         for tax_node in tax_nodes:
             amount = float(tax_node.text)
             domain = [
-                ('company_id', '=', journal.company_id.id),
+                ('company_id', '=', invoice_line.company_id.id),
                 ('amount_type', '=', 'percent'),
-                ('type_tax_use', '=', journal.type),
+                ('type_tax_use', '=', invoice_line.move_id.journal_id.type),
                 ('amount', '=', amount),
             ]
             tax_excl = self.env['account.tax'].search(domain + [('price_include', '=', False)], limit=1)
@@ -583,17 +584,17 @@ class AccountEdiCommon(models.AbstractModel):
                 inv_line_vals['taxes'].append(tax_incl.id)
                 inv_line_vals['price_unit'] *= (1 + tax_incl.amount / 100)
             else:
-                logs.append(_("Could not retrieve the tax: %s %% for line '%s'.", amount, invoice_line_form.name))
+                logs.append(_("Could not retrieve the tax: %s %% for line '%s'.", amount, invoice_line.name))
         # Set the values on the line_form
-        invoice_line_form.quantity = inv_line_vals['quantity']
+        invoice_line.quantity = inv_line_vals['quantity']
         if inv_line_vals.get('product_uom_id'):
-            invoice_line_form.product_uom_id = inv_line_vals['product_uom_id']
+            invoice_line.product_uom_id = inv_line_vals['product_uom_id']
         else:
             logs.append(
-                _("Could not retrieve the unit of measure for line with label '%s'.", invoice_line_form.name))
-        invoice_line_form.price_unit = inv_line_vals['price_unit']
-        invoice_line_form.discount = inv_line_vals['discount']
-        invoice_line_form.tax_ids = inv_line_vals['taxes']
+                _("Could not retrieve the unit of measure for line with label '%s'.", invoice_line.name))
+        invoice_line.price_unit = inv_line_vals['price_unit']
+        invoice_line.discount = inv_line_vals['discount']
+        invoice_line.tax_ids = inv_line_vals['taxes']
         return logs
 
     # -------------------------------------------------------------------------
@@ -610,9 +611,7 @@ class AccountEdiCommon(models.AbstractModel):
         elif invoice.move_type == 'out_refund':
             ecosio_format = ecosio_formats['credit_note']
         else:
-            invoice.with_context(no_new_invoice=True).message_post(
-                body="ECOSIO: could not validate xml, formats only exist for invoice or credit notes"
-            )
+            invoice.message_post(body="ECOSIO: could not validate xml, formats only exist for invoice or credit notes")
             return
         if not ecosio_format:
             return
@@ -635,11 +634,9 @@ class AccountEdiCommon(models.AbstractModel):
                         "<li><font style='color:Tomato;'><strong>" + detail['errorText'] + "</strong></font></li>")
 
         if errors_cnt == 0:
-            invoice.with_context(no_new_invoice=True).message_post(
-                body=f"<font style='color:Green;'><strong>ECOSIO: All clear for format {ecosio_format}!</strong></font>"
-            )
+            invoice.message_post(body=f"<font style='color:Green;'><strong>ECOSIO: All clear for format {ecosio_format}!</strong></font>")
         else:
-            invoice.with_context(no_new_invoice=True).message_post(
+            invoice.message_post(
                 body=f"<font style='color:Tomato;'><strong>ECOSIO ERRORS/WARNINGS for format {ecosio_format}</strong></font>: <ul> "
                      + "\n".join(report) + " </ul>"
             )
