--- addons/account_edi_ubl_cii/models/account_edi_format.py
@@ -1,83 +1,0 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from odoo import models, fields, _
-
-import logging
-
-_logger = logging.getLogger(__name__)
-
-FORMAT_CODES = [
-    'facturx_1_0_05',
-    'ubl_bis3',
-    'ubl_de',
-    'nlcius_1',
-    'efff_1',
-    'ubl_2_1',
-    'ubl_a_nz',
-]
-
-class AccountEdiFormat(models.Model):
-    _inherit = 'account.edi.format'
-
-    ####################################################
-    # Helpers
-    ####################################################
-
-    def _infer_xml_builder_from_tree(self, tree):
-        self.ensure_one()
-        ubl_version = tree.find('{*}UBLVersionID')
-        customization_id = tree.find('{*}CustomizationID')
-        if tree.tag == '{urn:un:unece:uncefact:data:standard:CrossIndustryInvoice:100}CrossIndustryInvoice':
-            return self.env['account.edi.xml.cii']
-        if customization_id is not None:
-            if 'xrechnung' in customization_id.text:
-                return self.env['account.edi.xml.ubl_de']
-            if customization_id.text == 'urn:cen.eu:en16931:2017#compliant#urn:fdc:peppol.eu:2017:poacc:billing:3.0':
-                return self.env['account.edi.xml.ubl_bis3']
-            if customization_id.text == 'urn:cen.eu:en16931:2017#compliant#urn:fdc:nen.nl:nlcius:v1.0':
-                return self.env['account.edi.xml.ubl_nl']
-        if ubl_version is not None:
-            if ubl_version.text == '2.0':
-                return self.env['account.edi.xml.ubl_20']
-            if ubl_version.text == '2.1':
-                return self.env['account.edi.xml.ubl_21']
-        return
-
-    ####################################################
-    # Import: Account.edi.format override
-    ####################################################
-
-    def _create_invoice_from_xml_tree(self, filename, tree, journal=None):
-        # EXTENDS account_edi
-        self.ensure_one()
-
-        if not self._is_ubl_cii_available(journal.company_id) and self.code != 'facturx_1_0_05':
-            return super()._create_invoice_from_xml_tree(filename, tree, journal=journal)
-
-        # infer the xml builder
-        invoice_xml_builder = self._infer_xml_builder_from_tree(tree)
-
-        if invoice_xml_builder is not None:
-            invoice = invoice_xml_builder._import_invoice(journal, filename, tree)
-            if invoice:
-                return invoice
-
-        return super()._create_invoice_from_xml_tree(filename, tree, journal=journal)
-
-    def _update_invoice_from_xml_tree(self, filename, tree, invoice):
-        # EXTENDS account_edi
-        self.ensure_one()
-
-        if not self._is_ubl_cii_available(invoice.company_id) and self.code != 'facturx_1_0_05':
-            return super()._update_invoice_from_xml_tree(filename, tree, invoice)
-
-        # infer the xml builder
-        invoice_xml_builder = self._infer_xml_builder_from_tree(tree)
-
-        if invoice_xml_builder is not None:
-            invoice = invoice_xml_builder._import_invoice(invoice.journal_id, filename, tree, invoice)
-            if invoice:
-                return invoice
-
-        return super()._update_invoice_from_xml_tree(filename, tree, invoice)
